// frontend/src/generated/CrmPropertyGroup.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CrmCatalogProperty } from "./CrmCatalogProperty";

/**
 * HubSpot のプロパティのグループ (画面の見出し) 1 つ
 */
export type CrmPropertyGroup = { 
/**
 * グループの内部名 (画面には出さない)
 */
name: string, label: string, 
/**
 * HubSpot の表示順
 */
properties: Array<CrmCatalogProperty>, };